2025 05 07 hiki

+++ academyfinal/src/main/java/com/hiki/academyfinal/dto/AttachmentDto.java
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 
 @Data @NoArgsConstructor @AllArgsConstructor @Builder
 public class AttachmentDto {
-	private int attachmentNo;		//시퀀스
+	private long attachmentNo;		//시퀀스
 	private String attachmentName;	//원본 파일명
 	private String attachmentType;	//파일 타입
 	private Long attachmentSize;		//바이트 단위 크기
